Sends an Email with System details
An expansion of the email script that sends an email with a timestamp and MAC address. Task scheduling it as an event would make it useful for logging USB use between certain systems. Wider application on autorun could add location details to track if plugged into networked device for security.